class文件解析增加一部分解析策略

+++ test/src/main/java/liufeng/jdk/jvm/compile/TestClass.java
@@ -3,7 +3,7 @@ package liufeng.jdk.jvm.compile;
 public class TestClass {
     private int m;
 
-    private long l = 3000L;
+    private double l = 3000d;
 
     public int inc() {
         return m + 1;
